final
All functions and features of the assignment are working as outlined in
the assignment document.
Checklist:
1. Current date appears, even with reset
2. cost field is updated as a user changes their quantity
3. Negatives and NaN in the quantity field throw errors, and add some
highlighting to show the user which value is invalid. resets to zero.
4. Shipping updates total based on which value is selected, throws error
if the user doesn't select an option
5. the form cannot submit unless:
a) at least 1 quantity field is filled out
and
b) a shipping choice is selected

It is calculating correctly based on testing the values in the
finalpage.jpg screenshot.